modules/module-06-etl-fundamentals/requirements.txt
# ETL Fundamentals - Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
sqlalchemy>=2.0.0

# Data validation
pydantic>=2.0.0
jsonschema>=4.17.0

# Testing
pytest>=7.3.0
pytest-cov>=4.0.0
faker>=18.0.0

# Logging
structlog>=23.1.0

# Database
sqlite3  # Included in Python standard library

# Optional: For advanced exercises
python-dotenv>=1.0.0
pyyaml>=6.0
